IS-1958: combine sykmeldinger with and without arbeidsgiver
